Delete Job node, favorite etc
